backend/requirements.txt
# Web framework
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0

# ASGI
uvicorn[standard]>=0.30,<1.0
gunicorn>=22.0,<23.0

# Database
psycopg2-binary>=2.9,<3.0

# Auth
djangorestframework-simplejwt>=5.3,<6.0

# WebSocket
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0

# Celery
celery[redis]>=5.4,<6.0

# Kafka
confluent-kafka>=2.5,<3.0

# HTTP client
httpx>=0.27,<1.0

# API docs
drf-spectacular>=0.27,<1.0

# Logging
structlog>=24.4,<25.0

# Config
django-environ>=0.11,<1.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
respx>=0.21,<1.0
